Made a few tweeks to error reporting.

+++ comex/src-mpi-pr/comex.c
@@ -5025,7 +5025,7 @@ STATIC xpmem_segid_t _shm_create(void **data, size_t size)
     comex_error("_shm_create mmap failed. size: ",(int)size);
   }
   segid = xpmem_make(ptr, size, XPMEM_PERMIT_MODE, (void *)06666);
-  if (segid == -1) {
+  if ((int64_t)segid == (int64_t)-1) {
     char buf[1024];
     sprintf(buf,"_shm_create xpmem_make failed. ptr: %p size: %ld",
         ptr,(int64_t)size);
